Add back common headers middleware
