+++ src/http.rs
@@ -4,7 +4,7 @@ use futures::future::{self, Future, BoxFuture, FutureResult};
 use futures::Stream;
 use handlers::return_to_relier;
 use hyper::{self, StatusCode, Error as HyperError};
-use hyper::header::{ContentType};
+use hyper::header::{ContentType, StrictTransportSecurity, CacheControl, CacheDirective};
 use hyper::server::{Request, Response, Service as HyperService};
 use hyper_tls::HttpsConnector;
 use std::collections::HashMap;
@@ -12,6 +12,13 @@ use std::error::Error;
 use std::sync::{Arc, Mutex};
 use tokio_core::reactor::Remote;
 use url::{Url, form_urlencoded};
+
+
+header! { (ContentSecurityPolicy, "Content-Security-Policy") => [String] }
+header! { (XContentSecurityPolicy, "X-Content-Security-Policy") => [String] }
+header! { (XContentTypeOptions, "X-Content-Type-Options") => [String] }
+header! { (XXSSProtection, "X-XSS-Protection") => [String] }
+header! { (XFrameOptions, "X-Frame-Options") => [String] }
 
 
 /// The default type of client we use for outgoing requests
@@ -62,7 +69,6 @@ impl HyperService for Service {
     type Error = HyperError;
     type Future = BoxFuture<Self::Response, Self::Error>;
 
-    // TODO Add headers
     fn call(&self, req: Request) -> Self::Future {
         info!("{} {}", req.method(), req.path());
 
@@ -74,6 +80,7 @@ impl HyperService for Service {
 
         (self.route)(&req)(req, ctx.clone())
             .or_else(|err| handle_error(ctx, err))
+            .map(|mut res| { set_headers(&mut res); res })
             .boxed()
     }
 }
@@ -153,6 +160,37 @@ fn handle_error(shared_ctx: ContextHandle, err: BrokerError) -> FutureResult<Res
 }
 
 
+/// Mutate a response to set common headers.
+fn set_headers(res: &mut Response) {
+    let headers = res.headers_mut();
+
+    // Specify a tight content security policy. We need to be able to POST
+    // redirect anywhere, and run our own scripts.
+    let csp = vec![
+        "sandbox allow-scripts allow-forms",
+        "default-src 'none'",
+        "script-src 'self'",
+        "style-src 'self'",
+        "form-action *",
+    ].join("; ");
+
+    headers.set(StrictTransportSecurity::excluding_subdomains(31_536_000u64));
+    headers.set(ContentSecurityPolicy(csp.clone()));
+    headers.set(XContentSecurityPolicy(csp));
+    headers.set(XContentTypeOptions("nosniff".to_string()));
+    headers.set(XXSSProtection("1; mode=block".to_string()));
+    headers.set(XFrameOptions("DENY".to_string()));
+
+    // Default to disable caching completely.
+    if !headers.has::<CacheControl>() {
+        headers.set(CacheControl(vec![
+            CacheDirective::NoCache,
+            CacheDirective::NoStore,
+        ]));
+    }
+}
+
+
 /// Parse a form-encoded string into a `HashMap`.
 pub fn parse_form_encoded(input: &[u8]) -> HashMap<String, String> {
     let mut map = HashMap::new();
@@ -179,4 +217,26 @@ pub fn parse_form_encoded_body(req: Request) -> BoxFuture<HashMap<String, String
         .map_err(|err| err.into())
         .map(|chunk| parse_form_encoded(&chunk))
         .boxed()
+}
+
+
+#[cfg(test)]
+mod tests {
+    use hyper::server::Response;
+    use super::set_headers;
+
+    #[test]
+    fn sets_expected_headers() {
+        let mut res = Response::new();
+        set_headers(&mut res);
+
+        let headers = res.headers();
+        assert!(headers.get_raw("Strict-Transport-Security").is_some());
+        assert!(headers.get_raw("Content-Security-Policy").is_some());
+        assert!(headers.get_raw("X-Content-Security-Policy").is_some());
+        assert!(headers.get_raw("X-Content-Type-Options").is_some());
+        assert!(headers.get_raw("X-XSS-Protection").is_some());
+        assert!(headers.get_raw("X-Frame-Options").is_some());
+        assert!(headers.get_raw("Cache-Control").is_some());
+    }
 }
